mrg/requirements.txt
tensorflow-gpu==1.12.0
hickle==3.4.3
tqdm==4.31.1
numpy==1.16.0
nltk==3.4.5 
pandas==0.25.1 
gensim==3.8.0  